+++ CHANGELOG.md
@@ -26,6 +26,7 @@
 - Rename `modytpe::field_param` to `modtype::non_static`.
 - Rename `{modtype, modtype::field_param, modtype::non_static}::DefaultModType` to `F` again.
 - `Cartridge::from_*` no longer return `Option`.
+- Remove type arguments from `ModType`s.
 
 ### Fixed